project structure changed: moved reducer and context in sep folder
