+++ src/App.js
@@ -6,7 +6,7 @@ import Home from "./components/Home";
 import Footer from "./components/Footer";
 import Toggle from "react-toggle";
 import "react-toggle/style.css";
-import { Context } from "./components/context";
+import { Context } from "./components/context/context";
 
 const todoCategories = [
   {
